# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ecommerce REST API built using Go Lang, PSQL, REST API following Clean Architecture. Hosted with Ngnix, AWS EC2 and RDS",
        "title": "Ecommerce REST API",
        "contact": {
            "name": "Nadeem Fahad",
            "url": "https://github.com/Nadeem1815",
            "email": "nadeemf408@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/user/buy_all": {
            "post": {
                "description": "User OrderProduct From Carts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "User BuyAllProduct from cart",
                "operationId": "user-order-product",
                "parameters": [
                    {
                        "description": "order Details",
                        "name": "order_details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_model.PlaceOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/user/cancelorder/{oderid}": {
            "put": {
                "description": "Update order for Admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Update Order for Admin",
                "operationId": "update-order",
                "parameters": [
                    {
                        "description": "orderid",
                        "name": "updating_details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_model.UpdateOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
        "/user/return": {
            "post": {
                "description": "ReturnRequest From  users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "ReturnRequest From  users",
                "operationId": "retrunreq-user",
                "parameters": [
                    {
                        "description": "Return details",
                        "name": "return_details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_model.RetrunRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/user/view": {
            "get": {
                "description": "Endpoint for getting all orders associated with a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Retrieves all orders of currently logged in user",
                "operationId": "view-all-orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/user/viewid/{order_id}": {
            "get": {
                "description": "Endpoint for getting  specific orders associated with a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Retrieves  ordersbyID of currently logged in user",
                "operationId": "view-orderID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "orderid",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_nadeem1815_premium-watch_pkg_utils_model.PlaceOrder": {
            "type": "object",
            "properties": {
                "payment_method_id": {
                    "type": "integer"
                },
                "shipping_address_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_nadeem1815_premium-watch_pkg_utils_model.RetrunRequest": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_nadeem1815_premium-watch_pkg_utils_model.UpdateOrder": {
            "type": "object",
            "properties": {
                "delivery_status_id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "status_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_nadeem1815_premium-watch_pkg_utils_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}